refactor: share focused child follow-up routing
